% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bounds.R
\name{get_bounds}
\alias{get_bounds}
\title{Get geoboundaries in a dataframe from GeoJSON with a code}
\usage{
get_bounds(code = "AUS")
}
\arguments{
\item{code}{3-letter code}
}
\value{
data frame with fields and geometry (see \link[wk:wkb]{wk::wkb})
}
\description{
Currently hardcoded to ADM1, see https://www.geoboundaries.org/api.html
}
\examples{
get_bounds("EGY")
}
